+++ src/main/java/com/esmile/appEsmile/service/impl/AppointmentService.java
@@ -1,11 +1,13 @@
 package com.esmile.appEsmile.service.impl;
 
 import com.esmile.appEsmile.entity.Appointment;
+import com.esmile.appEsmile.exception.ResourceNotFoundException;
 import com.esmile.appEsmile.repository.IAppointmentRepository;
 import com.esmile.appEsmile.service.IService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -22,8 +24,9 @@ public class AppointmentService implements IService<Appointment> {
     }
 
     @Override
-    public Optional<Appointment> get(Long id) {
-        return appointmentRepository.findById(id) ;
+    public Optional<Appointment> get(Long id) throws ResourceNotFoundException {
+
+        return Optional.ofNullable(appointmentRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("")));
     }
 
     @Override
@@ -37,8 +40,9 @@ public class AppointmentService implements IService<Appointment> {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws ResourceNotFoundException {
+        if (appointmentRepository.findById(id).isEmpty())
+            throw new ResourceNotFoundException("");
         appointmentRepository.deleteById(id);
-        //return null;
     }
 }
