+++ src/main/java/com/esmile/appEsmile/service/impl/DentistService.java
@@ -1,16 +1,18 @@
 package com.esmile.appEsmile.service.impl;
 
 import com.esmile.appEsmile.entity.Dentist;
+import com.esmile.appEsmile.exception.ResourceNotFoundException;
 import com.esmile.appEsmile.repository.IDentistRepository;
 import com.esmile.appEsmile.service.IService;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class DentistService implements IService<Dentist> {
 
-//    @Autowired
+    //    @Autowired
     private final IDentistRepository dentistRepository;
 
     public DentistService(IDentistRepository dentistRepository) {
@@ -38,8 +40,9 @@ public class DentistService implements IService<Dentist> {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws ResourceNotFoundException {
+        if (dentistRepository.findById(id).isEmpty())
+            throw new ResourceNotFoundException("");
         dentistRepository.deleteById(id);
-        //return null;
     }
 }
